# Use a single condvar/mutex pair for synchronization