--- a/Cargo.toml
+++ b/Cargo.toml
@@ -20,7 +20,6 @@ harness = false
 
 [dependencies]
 crossbeam-deque = "0.8"
-crossbeam-utils = "0.8"
 log = "0.4"
 profiling = "1"